# docs/swagger.yaml
definitions:
  dto.SubscriptionCreateRequest:
    properties:
      end_date:
        example: 05-2026
        type: string
      price:
        example: 599
        minimum: 0
        type: integer
      service_name:
        example: Netflix
        maxLength: 100
        type: string
      start_date:
        example: 01-2026
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.SubscriptionUpdateRequest:
    properties:
      end_date:
        example: 05-2026
        type: string
      price:
        example: 699
        minimum: 0
        type: integer
      service_name:
        example: Netflix Premium
        maxLength: 100
        type: string
      start_date:
        example: 01-2026
        type: string
    type: object
  handlers.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      err:
        $ref: '#/definitions/handlers.Error'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Это API для агрегации данных об онлайн-подписках пользователей.
  title: Сервис онлайн-подписок
  version: "1.0"
paths:
  /subscription:
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку для пользователя
      parameters:
      - description: Данные для создания подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionCreateRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать новую подписку
      tags:
      - subscription
  /subscription/:
    get:
      consumes:
      - application/json
      description: Возвращает список всех подписок
      parameters:
      - description: Номер страницы
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      - description: Количество записей на странице
        in: query
        minimum: 0
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить все подписки
      tags:
      - subscription
  /subscription/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку по её ID и возвращает информацию об удалённой
        подписке
      parameters:
      - description: ID подписки для удаления
        in: path
        minimum: 0
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Удалить подписку по ID
      tags:
      - subscription
    get:
      consumes:
      - application/json
      description: Возвращает информацию о конкретной подписке по её ID
      parameters:
      - description: ID подписки
        in: path
        minimum: 0
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить подписку по ID
      tags:
      - subscription
    patch:
      consumes:
      - application/json
      description: Обновляет информацию о существующей подписке
      parameters:
      - description: ID подписки для обновления
        in: path
        minimum: 0
        name: id
        required: true
        type: integer
      - description: Данные для обновления подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionUpdateRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscription
  /subscription/price:
    get:
      consumes:
      - application/json
      description: Рассчитывает общую стоимость подписок по заданным фильтрам (пользователь,
        сервис, период)
      parameters:
      - description: UUID пользователя для фильтрации
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Название сервиса для фильтрации
        in: query
        name: service_name
        type: string
      - description: Начальная дата для подсчета суммы
        in: query
        name: start_date
        required: true
        type: string
      - description: Конечная дата для подсчета суммы
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить сумму подписок по фильтрам
      tags:
      - subscription
  /subscription/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех подписок пользователя по его ID
      parameters:
      - description: Номер страницы
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      - description: Количество записей на странице
        in: query
        minimum: 0
        name: limit
        required: true
        type: integer
      - description: UUID пользователя
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Неверный формат UUID или параметры запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить подписки пользователя
      tags:
      - subscription
swagger: "2.0"